+++ filter_iterators_row.cpp
@@ -140,7 +140,7 @@ private:
 
 int main() {
   table<int,int,int,int> values;
-  values.reserve(end_data() - begin_data());
+  values.reserve(static_cast<size_t>(end_data() - begin_data()));
   for (const auto* p = begin_data(); p != end_data(); ++p) {
     values.insert(p->x, p->y, p->z, p->d);
   }
