+++ templated.cpp
@@ -69,7 +69,7 @@ void drop_if(V& v, P p)
 int main()
 {
   table<int,int,int,int> values;
-  values.reserve(end_data() - begin_data());
+  values.reserve(static_cast<size_t>(end_data() - begin_data()));
   for (const auto* p = begin_data(); p != end_data(); ++p) {
     values.push_back(p->x, p->y, p->z, p->d);
   }
